7.1.14 Refactor: splits actions
