--- src/store/api-actions.js
@@ -1,9 +1,0 @@
-import {loadOffers} from "./action";
-import {adaptOffer} from "../utils";
-
-export const fetchOffersList = () => (dispatch, _getState, api) => (
-  api.get(`/hotels`)
-    .then(({data}) => dispatch(loadOffers(
-        data.map((it) => adaptOffer(it))
-    )))
-);
